requirements.txt
aniso8601==9.0.1
beautifulsoup4==4.11.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.3
cssselect==1.2.0
cymem==2.0.7
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
lxml==4.9.1
MarkupSafe==2.1.1
murmurhash==1.0.9
newspaper3k==0.2.8
nltk==3.7
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pathy==0.8.1
Pillow==9.3.0
pip==22.2.2
preshed==3.0.8
pydantic==1.10.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
setuptools==65.5.0
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.2.post1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
thinc==8.1.5
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.12
wasabi==0.10.1
Werkzeug==2.2.2
wheel==0.37.1
zipp==3.10.0
